Decouple the forwarding plane dependencies
