+++ test/integration/basic_excluded_prefixes_test.go
@@ -40,7 +40,7 @@ func TestExcludePrefixCheck(t *testing.T) {
 	nodes, err := kubetest.SetupNodesConfig(k8s, nodesCount, defaultTimeout, []*pods.NSMgrPodConfig{
 		{
 			Variables:          variables,
-			DataplaneVariables: kubetest.DefaultDataplaneVariables(),
+			DataplaneVariables: kubetest.DefaultDataplaneVariables(k8s.GetForwardingPlane()),
 			Namespace:          k8s.GetK8sNamespace(),
 		},
 	}, k8s.GetK8sNamespace())
