+++ test/integration/basic_nsmd_deploy_test.go
@@ -16,14 +16,14 @@ import (
 )
 
 func TestNSMgrDdataplaneDeploy(t *testing.T) {
-	testNSMgrDdataplaneDeploy(t, pods.NSMgrPod, pods.VPPDataplanePod)
+	testNSMgrDdataplaneDeploy(t, pods.NSMgrPod, pods.ForwardingPlane)
 }
 
 func TestNSMgrDdataplaneDeployLiveCheck(t *testing.T) {
-	testNSMgrDdataplaneDeploy(t, pods.NSMgrPodLiveCheck, pods.VPPDataplanePodLiveCheck)
+	testNSMgrDdataplaneDeploy(t, pods.NSMgrPodLiveCheck, pods.ForwardingPlaneWithLiveCheck)
 }
 
-func testNSMgrDdataplaneDeploy(t *testing.T, nsmdPodFactory func(string, *v1.Node, string) *v1.Pod, dataplanePodFactory func(string, *v1.Node) *v1.Pod) {
+func testNSMgrDdataplaneDeploy(t *testing.T, nsmdPodFactory func(string, *v1.Node, string) *v1.Pod, dataplanePodFactory func(string, *v1.Node, string) *v1.Pod) {
 	RegisterTestingT(t)
 
 	if testing.Short() {
