+++ test/integration/recover_death_handling_test.go
@@ -23,7 +23,7 @@ var nseNoHeal = &pods.NSMgrPodConfig{
 		nsm.NsmdHealDSTWaitTimeout:   "1",    // 1 second
 		nsm.NsmdHealEnabled:          "true",
 	},
-	DataplaneVariables: kubetest.DefaultDataplaneVariables(),
+	DataplaneVariables: kubetest.DefaultDataplaneVariables(k8s.GetForwardingPlane()),
 }
 
 func TestNSCDiesSingleNode(t *testing.T) {
